news/requirements.txt
currentsapi
python-dotenv
pyspark